Moved util methods into drools-utils
