+++ drools-model/drools-model-codegen/pom.xml
@@ -40,6 +40,10 @@
     <dependencies>
         <dependency>
             <groupId>org.drools</groupId>
+            <artifactId>drools-util</artifactId>
+        </dependency>    
+        <dependency>
+            <groupId>org.drools</groupId>
             <artifactId>drools-codegen-common</artifactId>
         </dependency>
         <dependency>
