+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/util/ClassUtil.java
@@ -56,50 +56,4 @@ public class ClassUtil {
         return ACCESSIBLE_PROPS_CACHE_INCLUDING_NON_GETTER.computeIfAbsent( clazz, PropertyReactivityUtil::getAccessiblePropertiesIncludingNonGetterValueMethod );
     }
 
-    public static Type boxTypePrimitive(Type type) {
-        if (type instanceof Class<?>) {
-            return MethodUtils.boxPrimitive((Class<?>)type);
-        } else {
-            return type;
-        }
-    }
-
-    public static boolean isAssignableFrom(Type from, Type to) {
-        Class<?> fromClass = toRawClass( from );
-        Class<?> toClass = toRawClass( to );
-        return fromClass.isAssignableFrom(toClass) || MethodUtils.areBoxingCompatible(fromClass, toClass);
-    }
-
-    public static Class<?> toNonPrimitiveType(Class<?> c) {
-        if (!c.isPrimitive()) return c;
-        if (c == int.class) return Integer.class;
-        if (c == long.class) return Long.class;
-        if (c == double.class) return Double.class;
-        if (c == float.class) return Float.class;
-        if (c == short.class) return Short.class;
-        if (c == byte.class) return Byte.class;
-        if (c == char.class) return Character.class;
-        if (c == boolean.class) return Boolean.class;
-        return c;
-    }
-
-    public static Class<?> toRawClass(Type type) {
-        if (type == null) {
-            return null;
-        }
-        if (type instanceof Class<?>) {
-            return ( Class ) type;
-        }
-        if (type instanceof ParameterizedType ) {
-            return toRawClass( (( ParameterizedType ) type).getRawType() );
-        }
-        if (type instanceof TypeVariable ) {
-            return Object.class;
-        }
-        throw new UnsupportedOperationException( "Unknown type " + type );
-    }
-
-    public static Type getTypeArgument(Type genericType, int index) {
-        return genericType instanceof ParameterizedType ? (( ParameterizedType ) genericType).getActualTypeArguments()[index] : Object.class;
-    }
 }
