+++ drools-model/drools-mvel-compiler/src/main/java/org/drools/mvelcompiler/util/TypeUtils.java
@@ -16,39 +16,15 @@
 
 package org.drools.mvelcompiler.util;
 
-import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
-import java.util.List;
-import java.util.Map;
 
 import com.github.javaparser.StaticJavaParser;
-import org.drools.mvelcompiler.MvelCompilerException;
-
-import static java.util.stream.Stream.of;
+import static org.drools.util.ClassUtils.classFromType;
 
 public class TypeUtils {
 
     private TypeUtils() {
 
-    }
-
-    public static boolean isCollection(Type t) {
-        return of(List.class, Map.class).anyMatch(cls -> {
-            Class<?> clazz = classFromType(t);
-            return cls.isAssignableFrom(clazz);
-        });
-    }
-
-    public static Class<?> classFromType(Type t) {
-        Class<?> clazz;
-        if(t instanceof Class<?>) {
-            clazz = (Class<?>) t;
-        } else if(t instanceof ParameterizedType) {
-            clazz = (Class<?>) ((ParameterizedType)t).getRawType();
-        } else {
-            throw new MvelCompilerException("Unable to parse type");
-        }
-        return clazz;
     }
 
     public static com.github.javaparser.ast.type.Type toJPType(Type t) {
